src/services/users/requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
deprecation==2.1.0
Django==5.2.7
django-redis==6.0.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jwcrypto==1.5.6
packaging==25.0
psycopg2-binary==2.9.11
pycparser==2.23
python-decouple==3.8
python-dotenv==1.1.1
python-keycloak==4.0.0
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
